Refactor notificationHeaderButton
